[#250] refactor: parsing service 분리
